Add token validation to actions
modified:   app/controllers/courses_controller.rb
modified:   app/controllers/grades_controller.rb
modified:   app/controllers/students_controller.rb
modified:   test/models/user_test.rb

+++ app/controllers/courses_controller.rb
@@ -1,11 +1,17 @@
 class CoursesController < ApplicationController
   protect_from_forgery with: :null_session
   def index
+    @user = user_validation
+    return unless @user
+
     course = Course.all.order(:id)
     render json: course, status: 200
   end
 
   def show
+    @user = user_validation
+    return unless @user
+
     course = Course.find_by(id: params[:id])
     if course
       render json: course, status: 200
@@ -15,6 +21,9 @@ class CoursesController < ApplicationController
   end
 
   def create
+    @user = user_validation
+    return unless @user
+
     course = Course.new
     course.name = params[:name]
     course.id = params[:id]
@@ -27,6 +36,9 @@ class CoursesController < ApplicationController
   end
 
   def update
+    @user = user_validation
+    return unless @user
+
     course = Course.find_by(id: params[:id])
     if course
       course.update!(course_params)
@@ -37,6 +49,9 @@ class CoursesController < ApplicationController
   end
 
   def destroy
+    @user = user_validation
+    return unless @user
+
     course = Course.find_by(id: params[:id])
     if course
       course.destroy
@@ -47,6 +62,15 @@ class CoursesController < ApplicationController
   end
 
   private
+
+  def user_validation
+    user, error, status = UsersHelper::Validator.valid_user_token?(request.headers['Authorization'])
+    unless user
+      render json: { 'error:': error }, status: status
+      return
+    end
+    user
+  end
 
   def course_params
     params.require(:course).permit(:id, :name)
